Apply CORS on validation errors (fixes #104)
